api/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
filelock==3.14.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
gensim==4.3.2
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
networkx==3.3
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.26.4
openai==1.35.13
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
playwright==1.42.0
pluggy==1.4.0
protobuf==4.25.3
psutil==5.9.8
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
py==1.11.0
pydantic==2.7.3
pydantic_core==2.18.4
pyee==11.0.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
pyparsing==3.1.2
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
pytest-xprocess==0.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scikit-surprise==1.1.4
scipy==1.14.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
text-unidecode==1.3
textblob==0.18.0.post0
tf_keras==2.16.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.41.2
twilio==9.0.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
